+++ README.md
@@ -1,13 +1,11 @@
 # Project Movies
 
-Replace this readme with your own information about your project.
-
-Start by briefly describing the assignment in a sentence or two. Keep it short and to the point.
+This project is for practicing fetching from API's with react, using react-router-dom for handling the pages.
 
 ## The problem
 
-Describe how you approached to problem, and what tools and techniques you used to solve it. How did you plan? What technologies did you use? If you had more time, what would be next?
+I started with the functionality, fetching all the data and getting the information needed on the pages, then I styled everything. I chose to make a stylesheet for every page or component to make them easier to reuse.
 
 ## View it live
 
-Every project should be deployed somewhere. Be sure to include the link to the deployed project so that the viewer can click around and see what it's all about.
+https://condescending-fermat-253ba9.netlify.app/
