+++ code/package.json
@@ -14,6 +14,7 @@
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
+    "react-spinners-css": "^1.2.2",
     "react-spring": "^8.0.27"
   },
   "scripts": {
