+++ code/src/components/Header.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './header.css';
+
+export const Header = () => {
+	return (
+		<header>
+			<h1>Welcome to the moooooviepage</h1>
+		</header>
+	);
+};
