+++ code/src/index.css
@@ -5,6 +5,8 @@ body {
 		sans-serif;
 	-webkit-font-smoothing: antialiased;
 	-moz-osx-font-smoothing: grayscale;
+	background-color: #000;
+	color: #fff;
 }
 
 .root {
