+++ code/src/pages/MovieList.js
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from 'react';
+import { Spinner } from 'react-spinners-css';
 
 import { MOVIEDB_URL } from '../components/url';
 import { Movie } from '../components/Movie';
@@ -7,15 +8,20 @@ import './movielist.css';
 
 export const MovieList = () => {
 	const [movies, setMovies] = useState([]);
+	const [loading, setLoading] = useState(true);
 
 	useEffect(() => {
 		fetch(MOVIEDB_URL)
 			.then((res) => res.json())
-			.then((json) => setMovies(json.results));
+			.then((json) => {
+				setLoading(false);
+				setMovies(json.results);
+			});
 	}, []);
 
 	return (
 		<section className="movielist__container">
+			{loading && <Spinner />}
 			{movies.map((movie) => (
 				<Movie key={movie.id} {...movie} />
 			))}
