+++ code/src/pages/PageNotFound.js
@@ -1,5 +1,18 @@
 import React from 'react';
 
+import { Button } from '../components/Button';
+
+import './pagenotfound.css';
+
 export const PageNotFound = () => {
-	return <h1>Something went wrong!</h1>;
+	return (
+		<div className="page-not-found">
+			<section className="page-not-found__error">
+				<h1>Something went wrong...</h1>
+				<h2>Go back to the movies page and pick another movie.</h2>
+
+				<Button />
+			</section>
+		</div>
+	);
 };
