+++ code/src/pages/pagenotfound.css
@@ -1,0 +1,6 @@
+.page-not-found {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	height: 400px;
+}
